// src/app/models/book.model.ts
export interface Book {
    id: string;
    volumeInfo: {
        title: string;
        authors?: string[];
        description?: string;
        imageLinks?: {
            thumbnail?: string;
        };
        publishedDate?: string;
    };
}

export interface BooksResponse {
    items: Book[];
}